add id check to AnswerController

+++ src/main/java/com/codecool/stackoverflowtw/controller/AnswerController.java
@@ -4,6 +4,7 @@ import com.codecool.stackoverflowtw.controller.dto.answer.AnswerDTO;
 import com.codecool.stackoverflowtw.controller.dto.answer.AnswerVoteDTO;
 import com.codecool.stackoverflowtw.controller.dto.answer.NewAnswerDTO;
 import com.codecool.stackoverflowtw.service.AnswerService;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -25,7 +26,10 @@ public class AnswerController {
   }
   
   @PostMapping ("/")
-  public int addNewAnswer(@RequestBody NewAnswerDTO newAnswerDTO) {
+  public int addNewAnswer(@RequestBody NewAnswerDTO newAnswerDTO, HttpServletResponse response) {
+    if (newAnswerDTO.userId() != getUserId(response)) {
+      return -1;
+    }
     return answerService.addNewAnswer(newAnswerDTO);
   }
   
@@ -35,12 +39,22 @@ public class AnswerController {
   }
   
   @PostMapping ("/votes")
-  public int addVoteToAnswer(@RequestBody AnswerVoteDTO answerVoteDTO) {
+  public int addVoteToAnswer(@RequestBody AnswerVoteDTO answerVoteDTO, HttpServletResponse response) {
+    if (answerVoteDTO.userId() != getUserId(response)) {
+      return -1;
+    }
     return answerService.addVoteToAnswer(answerVoteDTO);
   }
   
   @DeleteMapping ("/votes/{aId}/{uId}")
-  public int deleteAnswerVote(@PathVariable int aId, @PathVariable int uId) {
+  public int deleteAnswerVote(@PathVariable int aId, @PathVariable int uId, HttpServletResponse response) {
+    if (uId != getUserId(response)) {
+      return -1;
+    }
     return answerService.deleteAnswerVote(aId, uId);
+  }
+  
+  private int getUserId(HttpServletResponse response) {
+    return Integer.parseInt(response.getHeader("user_id"));
   }
 }
